requirements.txt
click==8.1.3
greenlet==2.0.2
mysql-connector-python==8.0.32
protobuf==3.20.3
python-dotenv==1.0.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
